+++ weaver/processes/wps1_process.py
@@ -136,7 +136,10 @@ class Wps1Process(WpsProcessInterface):
         LOGGER.debug("Execute WPS-1 provider: [%s]", self.provider)
         LOGGER.debug("Execute WPS-1 process: [%s]", self.process)
         try:
-            self.wps_provider = get_wps_client(self.provider, headers=self.cookies)
+            headers = {}
+            headers.update(self.get_auth_cookies())
+            headers.update(self.get_auth_headers())
+            self.wps_provider = get_wps_client(self.provider, headers=headers)
             raise_on_xml_exception(self.wps_provider._capabilities)  # noqa: W0212
         except Exception as ex:
             raise OWSNoApplicableCode("Failed to retrieve WPS capabilities. Error: [{}].".format(str(ex)))